requirements-docker.txt
# CPU-only AI requirements for Docker
python-dotenv==1.0.0
imapclient==3.0.1
pdfplumber==0.10.3
PyPDF2==3.0.1
requests==2.31.0

# AI/ML libraries - CPU only (no CUDA)
torch==2.1.2+cpu -f https://download.pytorch.org/whl/torch_stable.html
transformers==4.36.2
tokenizers==0.15.0
numpy==1.24.4
pillow==10.1.0

# Additional dependencies for transformers
tqdm
pyyaml
filelock
huggingface-hub
safetensors
regex
packaging
